pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
